+++ lib/internal/modules/esm/translators.js
@@ -18,16 +18,6 @@ const {
   globalThis: { WebAssembly },
 } = primordials;
 
-/** @type {import('internal/util/types')} */
-let _TYPES = null;
-/**
- * Lazily loads and returns the internal/util/types module.
- */
-function lazyTypes() {
-  if (_TYPES !== null) { return _TYPES; }
-  return _TYPES = require('internal/util/types');
-}
-
 const {
   compileFunctionForCJSLoader,
 } = internalBinding('contextify');
@@ -37,7 +27,9 @@ const assert = require('internal/assert');
 const { readFileSync } = require('fs');
 const { dirname, extname, isAbsolute } = require('path');
 const {
+  assertBufferSource,
   loadBuiltinModule,
+  stringify,
   stripTypeScriptTypes,
   stripBOM,
   urlToFilename,
@@ -57,7 +49,6 @@ let debug = require('internal/util/debuglog').debuglog('esm', (fn) => {
 const { emitExperimentalWarning, kEmptyObject, setOwnProperty, isWindows } = require('internal/util');
 const {
   ERR_UNKNOWN_BUILTIN_MODULE,
-  ERR_INVALID_RETURN_PROPERTY_VALUE,
 } = require('internal/errors').codes;
 const { maybeCacheSourceMap } = require('internal/source_map/source_map_cache');
 const moduleWrap = internalBinding('module_wrap');
@@ -106,44 +97,6 @@ function initCJSParseSync() {
 
 const translators = new SafeMap();
 exports.translators = translators;
-
-let DECODER = null;
-/**
- * Asserts that the given body is a buffer source (either a string, array buffer, or typed array).
- * Throws an error if the body is not a buffer source.
- * @param {string | ArrayBufferView | ArrayBuffer} body - The body to check.
- * @param {boolean} allowString - Whether or not to allow a string as a valid buffer source.
- * @param {string} hookName - The name of the hook being called.
- * @throws {ERR_INVALID_RETURN_PROPERTY_VALUE} If the body is not a buffer source.
- */
-function assertBufferSource(body, allowString, hookName) {
-  if (allowString && typeof body === 'string') {
-    return;
-  }
-  const { isArrayBufferView, isAnyArrayBuffer } = lazyTypes();
-  if (isArrayBufferView(body) || isAnyArrayBuffer(body)) {
-    return;
-  }
-  throw new ERR_INVALID_RETURN_PROPERTY_VALUE(
-    `${allowString ? 'string, ' : ''}array buffer, or typed array`,
-    hookName,
-    'source',
-    body,
-  );
-}
-
-/**
- * Converts a buffer or buffer-like object to a string.
- * @param {string | ArrayBuffer | ArrayBufferView} body - The buffer or buffer-like object to convert to a string.
- * @returns {string} The resulting string.
- */
-function stringify(body) {
-  if (typeof body === 'string') { return body; }
-  assertBufferSource(body, false, 'load');
-  const { TextDecoder } = require('internal/encoding');
-  DECODER = DECODER === null ? new TextDecoder() : DECODER;
-  return DECODER.decode(body);
-}
 
 /**
  * Converts a URL to a file path if the URL protocol is 'file:'.
