+++ lib/internal/modules/helpers.js
@@ -15,6 +15,7 @@ const {
 } = primordials;
 const {
   ERR_INVALID_ARG_TYPE,
+  ERR_INVALID_RETURN_PROPERTY_VALUE,
 } = require('internal/errors').codes;
 const { BuiltinModule } = require('internal/bootstrap/realm');
 
@@ -384,8 +385,59 @@ function isUnderNodeModules(filename) {
   return ArrayPrototypeIncludes(splitPath, 'node_modules');
 }
 
+/** @type {import('internal/util/types')} */
+let _TYPES = null;
+/**
+ * Lazily loads and returns the internal/util/types module.
+ */
+function lazyTypes() {
+  if (_TYPES !== null) { return _TYPES; }
+  return _TYPES = require('internal/util/types');
+}
+
+
+/**
+ * Asserts that the given body is a buffer source (either a string, array buffer, or typed array).
+ * Throws an error if the body is not a buffer source.
+ * @param {string | ArrayBufferView | ArrayBuffer} body - The body to check.
+ * @param {boolean} allowString - Whether or not to allow a string as a valid buffer source.
+ * @param {string} hookName - The name of the hook being called.
+ * @throws {ERR_INVALID_RETURN_PROPERTY_VALUE} If the body is not a buffer source.
+ */
+function assertBufferSource(body, allowString, hookName) {
+  if (allowString && typeof body === 'string') {
+    return;
+  }
+  const { isArrayBufferView, isAnyArrayBuffer } = lazyTypes();
+  if (isArrayBufferView(body) || isAnyArrayBuffer(body)) {
+    return;
+  }
+  throw new ERR_INVALID_RETURN_PROPERTY_VALUE(
+    `${allowString ? 'string, ' : ''}array buffer, or typed array`,
+    hookName,
+    'source',
+    body,
+  );
+}
+
+let DECODER = null;
+
+/**
+ * Converts a buffer or buffer-like object to a string.
+ * @param {string | ArrayBuffer | ArrayBufferView} body - The buffer or buffer-like object to convert to a string.
+ * @returns {string} The resulting string.
+ */
+function stringify(body) {
+  if (typeof body === 'string') { return body; }
+  assertBufferSource(body, false, 'load');
+  const { TextDecoder } = require('internal/encoding');
+  DECODER = DECODER === null ? new TextDecoder() : DECODER;
+  return DECODER.decode(body);
+}
+
 module.exports = {
   addBuiltinLibsToObject,
+  assertBufferSource,
   getBuiltinModule,
   getCjsConditions,
   initializeCjsConditions,
@@ -394,6 +446,7 @@ module.exports = {
   makeRequireFunction,
   normalizeReferrerURL,
   stripTypeScriptTypes,
+  stringify,
   stripBOM,
   toRealPath,
   hasStartedUserCJSExecution() {
