+++ test/fixtures/typescript/ts/issue-54457.mjs
@@ -1,0 +1,4 @@
+// https://github.com/nodejs/node/issues/54457
+const { text } = await import('./test-require.ts');
+
+console.log(text);
