+++ test/fixtures/typescript/ts/test-require.ts
@@ -1,0 +1,7 @@
+const util = require('node:util');
+
+const text = 'Hello, TypeScript!';
+
+module.exports = {
+    text
+};
